requirements.txt
python-dotenv>=0.5
wtforms-json==0.2.10
weblib==0.1.20
bcrypt==2.0.0
blinker==1.4
cffi==1.7.0
click==6.6
Flask==0.11.1
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.1
Flask-User==0.6.8
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
passlib==1.6.5
pip==8.1.2
pycparser==2.14
pycrypto==2.6.1
setuptools==23.1.0
six==1.10.0
SQLAlchemy==1.0.13
Werkzeug==0.11.10
wheel==0.29.0
WTForms==2.1
